actions: Split out zerver.actions.realm_logo.
