+++ zerver/lib/actions.py
@@ -109,7 +109,6 @@ from zerver.lib.notification_data import UserMessageNotificationsData, get_user_
 from zerver.lib.pysa import mark_sanitized
 from zerver.lib.queue import queue_json_publish
 from zerver.lib.realm_icon import realm_icon_url
-from zerver.lib.realm_logo import get_realm_logo_data
 from zerver.lib.recipient_users import recipient_for_user_profiles
 from zerver.lib.retention import move_messages_to_archive
 from zerver.lib.send_email import (
@@ -4399,36 +4398,6 @@ def do_change_icon_source(
             active_user_ids(realm.id),
         )
     )
-
-
-@transaction.atomic(durable=True)
-def do_change_logo_source(
-    realm: Realm, logo_source: str, night: bool, *, acting_user: Optional[UserProfile]
-) -> None:
-    if not night:
-        realm.logo_source = logo_source
-        realm.logo_version += 1
-        realm.save(update_fields=["logo_source", "logo_version"])
-
-    else:
-        realm.night_logo_source = logo_source
-        realm.night_logo_version += 1
-        realm.save(update_fields=["night_logo_source", "night_logo_version"])
-
-    RealmAuditLog.objects.create(
-        event_type=RealmAuditLog.REALM_LOGO_CHANGED,
-        realm=realm,
-        event_time=timezone_now(),
-        acting_user=acting_user,
-    )
-
-    event = dict(
-        type="realm",
-        op="update_dict",
-        property="night_logo" if night else "logo",
-        data=get_realm_logo_data(realm, night),
-    )
-    transaction.on_commit(lambda: send_event(realm, event, active_user_ids(realm.id)))
 
 
 @transaction.atomic(durable=True)
